requirements.txt
# Core libraries
numpy>=1.19.0
pandas>=1.1.0
matplotlib>=3.3.0
scikit-learn>=0.23.0
pillow>=8.0.0
tqdm>=4.48.0
seaborn>=0.11.0

# PyTorch (CPU version, for GPU add cuda support)
torch>=1.9.0
torchvision>=0.10.0

# Deep learning utilities
timm>=0.5.4
albumentations>=1.0.0
